feat: adding endpoints

+++ managment/src/main/resources/db/migration/V1__create-table-employee.sql
@@ -2,7 +2,7 @@ CREATE TABLE employee(
     id TEXT PRIMARY KEY UNIQUE NOT NULL,
     nome TEXT NOT NULL,
     data_nascimento DATE NOT NULL,
-    salario FLOAT NOT NULL,
+    salario NUMERIC(10,2) NOT NULL,
     funcao TEXT NOT NULL
 );
